build: ensure that the correct Python version is used
On some platforms (Mac OS and Windows as per
https://github.com/actions/setup-python/issues/121#issuecomment-777748504),
CMake would not use the correct version of Python. Since we already have
a correct version of Python running the wrapper around CMake, and since
that thing was already trying to pass the interpreter to FindPython,
let's make sure that a correctly cased variable name is used.

Tested on GitHub Actions under Mac OS 12.4, and that thing appears to
work.