Always run the test suite when building the wheel
Unfortunately there's no Python-wide way of saying "this is how you run
tests", and no Python-wide way of saying "please run the tests when
building the wheel". If you ask me, this is obviously wrong because
tests are usually written for a reason, and not just for the lulz, and
it is a good thing to know that the build did not produce utter crap.

Since there's no Python-ish way of achieving this, let's give up and
hardcode everything. That way the test suite will be run as a part of
each build.

In the CI we're still taking the wheel in another VM and executing the
test suite again to make sure that whatever wheel magic which had to
happen happened correctly. That, however, is a different test from "is
the build OK" (e.g., `auditwheel repair`), and there's still point in
having an "early warning" should the build produce a non-working result.
Also, it (neatly) prevents a possible scenario where a wheel was left
unchecked by a second, testing CI job.

Bug: https://github.com/pypa/setuptools/issues/1684